+++ src/main/java/com/prowidesoftware/swift/model/field/StructuredNarrative.java
@@ -36,12 +36,12 @@ import org.apache.commons.lang3.StringUtils;
  */
 public class StructuredNarrative {
     private final List<String> narrativeFragments = new ArrayList<>();
+    private final transient List<NarrativeFragment> narrativeFragmentsDetail = new ArrayList<>();
     private final List<String> narrativeSupplementFragments = new ArrayList<>();
     private String codeword;
     private String currency;
     private BigDecimal amount;
     private String country;
-
     private String bankCode;
 
     /**
@@ -136,10 +136,34 @@ public class StructuredNarrative {
     }
 
     /**
-     * @see #getNarrativeFragments()
+     * Returns the list of text fragments in the narrative including the line index and line length,
+     * fragments are segments of the text that is wrapped in lines.
+     * If the narrative is a single string in a single line, the list will contain a single element.
+     * To get the narrative as a simple joined string use {@link #getNarrative(String)}
+     *
+     * @return the narrative fragments or an empty list if the narrative does not have any text
+     */
+    public List<NarrativeFragment> getNarrativeFragmentsDetail() {
+        return narrativeFragmentsDetail;
+    }
+
+    /**
+     * @see #addNarrativeFragment(String, int, int)
      */
     StructuredNarrative addNarrativeFragment(String narrativeFragment) {
+        return addNarrativeFragment(narrativeFragment, 0, 0);
+    }
+
+    /**
+     * Adds a fragment indicating the line index (1 based) and the line length.
+     *
+     * @param narrativeFragment text of the fragment
+     * @param lineIndex complete narrative line index
+     * @param lineLength complete line length
+     */
+    StructuredNarrative addNarrativeFragment(String narrativeFragment, int lineIndex, int lineLength) {
         this.narrativeFragments.add(narrativeFragment);
+        this.narrativeFragmentsDetail.add(new NarrativeFragment(narrativeFragment, lineIndex, lineLength));
         return this;
     }
 
@@ -183,7 +207,7 @@ public class StructuredNarrative {
     public String getNarrative(String delimiter) {
         if (!this.narrativeFragments.isEmpty()) {
             String s = delimiter != null ? delimiter : "";
-            return String.join(s, this.narrativeFragments);
+            return String.join(s, this.getNarrativeFragments());
         }
         return null;
     }
@@ -216,7 +240,7 @@ public class StructuredNarrative {
     }
 
     /**
-     * @return true if non of the narrative fields are set
+     * @return true if none of the narrative fields are set
      */
     public boolean isEmpty() {
         return this.codeword == null