+++ src/test/java/com/prowidesoftware/swift/model/mt/mt0xx/MT020Test.java
@@ -34,7 +34,7 @@ public class MT020Test {
 
     private final String sample3 = "{1:F01VNDZBET2AXXX0000000000}{2:I020DYDYXXXXXXXXN}{4:"
             + "{102:VNDZBET2AXXX}"
-            + "{253:050719MVNDZBET2AXXX0181000391}}";
+            + "{253:050719VNDZBET2AXXX0181000391}}";
 
     private final String sample4 = "{1:F01VNDZBET2AXXX0000000000}{2:I020DYDYXXXXXXXXN}{4:"
             + "{102:VNDZBET2AXXX}"
@@ -91,7 +91,7 @@ public class MT020Test {
         assertEquals("VNDZBET2AXXX", m.getField102().getValue());
 
         assertNotNull(m.getField253());
-        assertEquals("050719MVNDZBET2AXXX0181000391", m.getField253().getValue());
+        assertEquals("050719VNDZBET2AXXX0181000391", m.getField253().getValue());
     }
 
     @Test
@@ -184,7 +184,7 @@ public class MT020Test {
         m.setSender("VNDZBET2AXXX");
         m.setReceiver("DYDYXXXXFXXX");
         m.append(new Field102("VNDZBET2AXXX"));
-        m.append(new Field253("050719MVNDZBET2AXXX0181000391"));
+        m.append(new Field253("050719VNDZBET2AXXX0181000391"));
 
         assertEquals(sample3, m.message());
     }